requirements.txt
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudinary==1.16.0
colour==0.1.5
cssselect2==0.2.1
decorator==4.4.0
dominate==2.3.5
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Colorpicker==0.9
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
mock==3.0.5
names==0.3.0
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
pycparser==2.19
python-dateutil==2.8.0
python-resize-image==1.1.19
pytz==2019.1
reportlab==3.5.23
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.4
subprocess.run==0.0.8
svglib==0.9.0
tinycss2==1.0.2
urllib3==1.25.3
validators==0.13.0
visitor==0.1.3
webcolors==1.3
webencodings==0.5.1
Werkzeug==0.15.4
WTForms==2.2.1
WTForms-Components==0.10.4
xlrd==1.2.0
xlwt==1.3.0
